% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_mane_mapping.R
\name{create_mane_mapping}
\alias{create_mane_mapping}
\title{create mane mapping data.frame for RNA to DNA}
\usage{
create_mane_mapping(df)
}
\arguments{
\item{df}{a data.frame with the column \code{bp}}
}
\description{
a function that creates a mapping between genomic coordinate systems.
}
