% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_orf.R
\name{describe_orf}
\alias{describe_orf}
\title{describe orf}
\usage{
describe_orf(x, how = c("kozak"), start = "ATG", stop = "(TAG)|(TAA)|(TGA)")
}
\arguments{
\item{x}{string, the sequence.}

\item{how}{vector/string. Takes "kozak" and "".}

\item{start}{start codon}

\item{stop}{stop codon}
}
\description{
describe all orfs in a sequence, by converting them into a data.frame
}
\details{
depends on \code{get_orf} and \code{get_kozak_strength}.
}
