% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_codon.R
\name{find_codon}
\alias{find_codon}
\title{find codon}
\usage{
find_codon(x, codon = "ATG")
}
\arguments{
\item{x}{string (dna sequence)}

\item{codon}{string, regex.}
}
\description{
use regex to return position of codon
}
