% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_orfs.R
\name{find_orfs}
\alias{find_orfs}
\title{find open reading frames}
\usage{
find_orfs(x, start = "ATG", stop = "(TAG)|(TAA)|(TGA)")
}
\arguments{
\item{x}{a string}

\item{start}{what start codon should be used}

\item{stop}{what stop codons should be used}
}
\description{
takes in a string and finds open reading frames (ORFs)
}
\note{
This function will use the first stoped codon matching a start codon.
The rest of the stop codons (in-frame) will not be considered for the corresponding start codon.
}
