% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_truncating_augs.R
\name{find_truncating_augs}
\alias{find_truncating_augs}
\title{find truncating AUGs}
\usage{
find_truncating_augs(x, start = "ATG", stop = "(TAG)|(TAA)|(TGA)")
}
\arguments{
\item{x}{string, sequence}

\item{start}{string, regex}

\item{stop}{string, regex}
}
\description{
find AUGs that are not in an open reading frame
}
