% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_altered_cds.R
\name{get_altered_cds}
\alias{get_altered_cds}
\title{get altered CDS}
\usage{
get_altered_cds(utr, cds)
}
\arguments{
\item{utr}{string. UTR sequence}

\item{cds}{string. CDS sequence}
}
\description{
a function to get CDS that are either 1) elongated or 2) truncated or 3) both.
}
