% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cap_to_start_len.R
\name{get_cap_to_start_len}
\alias{get_cap_to_start_len}
\title{get length to start position}
\usage{
get_cap_to_start_len(x, f = get_orf)
}
\arguments{
\item{x}{string. a sequence}

\item{f}{a function that returns a named list with names indicating starts.
E.g. \code{f = get_orf} will return all open reading frames, and this function
will return the position of the first start codon in an ORF.}
}
\description{
get length to start position
}
