% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cds_proximity.R
\name{get_cds_proximity}
\alias{get_cds_proximity}
\title{get proximity to CDS}
\usage{
get_cds_proximity(x, fun = function(x) get_orf(x, share_stops = F))
}
\description{
finds the distance to the CDS (assuming that it will be the end of x)
}
