% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_kozak_strength.R
\name{get_kozak_strength}
\alias{get_kozak_strength}
\title{get kozak strength}
\usage{
get_kozak_strength(x, only_orf = F)
}
\arguments{
\item{x}{sequence}
}
\description{
get kozak strength (Assumes AUG as start codon)
}
\note{
currently this only check for kozaks in open reading frames.
}
