% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cds_proximity.R
\name{get_leader_proximity}
\alias{get_leader_proximity}
\title{get proximity to leader/cap}
\usage{
get_leader_proximity(x, fun = function(x) get_orf(x, share_stops = F))
}
\arguments{
\item{x}{sequence}

\item{fun}{what function should be appliedd}
}
\description{
get proximity to leader/cap
}
