% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_oorf_kozak.R
\name{get_oorf_kozak}
\alias{get_oorf_kozak}
\title{get overlapping open reading frames kozak strength}
\usage{
get_oorf_kozak(utr, cds, f = get_altered_cds)
}
\arguments{
\item{utr}{string. UTR sequence.}

\item{cds}{string. CDS sequence}

\item{f}{function to be used that takes two arguments, 
Ideally, it should be \code{get_altered_cds} or \code{get_oorf}.}
}
\value{
a value of kozak strength (3 strong, 2 moderate and 1 weak)
}
\description{
find kozak strength of overlapping reading frames.
}
