% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_bar.R
\name{gg_bar}
\alias{gg_bar}
\title{create a colored barplot}
\usage{
gg_bar(df, bonf, colors = c("firebrick2", "tomato1", "black"))
}
\arguments{
\item{df}{a data.frame with list_name and pvalue in data}

\item{bonf}{bonferroni correction threshold}

\item{colors}{colors used not significant, significant and bonferroni significant entries.}
}
\description{
creates a colored barplot based on bonferroni correction.
}
\concept{plotting}
