% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_mane.R
\name{map_mane}
\alias{map_mane}
\title{map geneomic coordiantes from RNA to DNA}
\usage{
map_mane(coords, bp, adjust_pos = -2)
}
\arguments{
\item{df}{a data.frame with the column \code{bp}.}
}
\description{
map geneomic coordiantes from RNA to DNA
}
