% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melt_tri.R
\name{melt_tri}
\alias{melt_tri}
\title{Melt iterable triangle}
\usage{
melt_tri(cols, rows = NULL, how = "upper", diagonal = FALSE)
}
\arguments{
\item{cols}{integer.}

\item{rows}{integer. Default is NULL, i.e. square matrix.}

\item{how}{string. "upper" or "lower".}

\item{diagonal}{boolean. Should diagonal be included?}
}
\description{
creates a long matrix that contains the paired coordinates
for non-redudant iteration.
}
