% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{pprint}
\alias{pprint}
\title{print to port}
\usage{
pprint(text, ...)
}
\arguments{
\item{text}{what text should be printed}

\item{...}{extra text}
}
\description{
log output to port
}
\details{
In the following order do:
* listen in terminal at: nc -l 4000
* set log.socket <- make.socket(port=4000)
* start script
}
