% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_workbook.R
\name{read_workbook}
\alias{read_workbook}
\title{read SaintExpress workbooks}
\usage{
read_workbook(
  workbook,
  sheets = NULL,
  rename_sheets = NULL,
  readxl_function = read_xlsx
)
}
\arguments{
\item{workbook}{string. Path to workbook.}

\item{sheets}{string(s). What sheets should be read? Default is reading all.}

\item{rename_sheets}{strings(s). Should sheetes be renamed?}

\item{readxl_function}{what function should be used to read in objects. Takes
the first two arguments as inputs.}
}
\description{
reads in a workbook using the readxl tidyverse library.
}
