% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_kozak.R
\name{select_kozak}
\alias{select_kozak}
\title{select kozak}
\usage{
select_kozak(x)
}
\description{
a sub function used to select among multiple sequences,
the kozak that has the highest kozak score (strongest). Ties will be 
resovled by selecting the one furthest away from the CDS start site.
}
