% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_expected_codons.R
\name{sim_expected_codons}
\alias{sim_expected_codons}
\title{simulate codon frequency}
\usage{
sim_expected_codons(
  seqs,
  codons = "ATG",
  k = 2,
  iter = 1000,
  verbose = T,
  parallel = F,
  cores = detectCores()
)
}
\arguments{
\item{seqs}{vector of sequences}

\item{k}{what k-mer frequency should be preserved?}

\item{iter}{how many simulations?}

\item{verbose}{print status?}

\item{parallel}{run parallel}

\item{cores}{how many cores should be used when in parallel?}

\item{codon}{what codon should be determined}
}
\description{
simulates codon frequency
}
