% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/som_prob_codons.R
\name{sim_prob_codons}
\alias{sim_prob_codons}
\title{simulate codon frequency}
\usage{
sim_prob_codons(seqs, codons = "ATG", k = 2, iter = 1000, verbose = T)
}
\arguments{
\item{seqs}{vector of sequences}

\item{codon}{what codon should be determined}
}
\description{
simulates codon frequency
}
