% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_seq.R
\name{sim_seq}
\alias{sim_seq}
\title{simulate codon frequency}
\usage{
sim_seq(seq, f, k = 2, iter = 1000, verbose = T)
}
\arguments{
\item{seq}{vector of sequences}

\item{f}{what function should be applied to simulate sequences? e.g. find_codon()}

\item{iter}{how many itereations}

\item{verbose}{printn status}
}
\description{
simulates codon frequency
}
