% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_kozak_strength.R
\name{get_kozak_strength}
\alias{get_kozak_strength}
\title{get kozak strength}
\usage{
get_kozak_strength(x, only_orf = F)
}
\arguments{
\item{x}{sequence}

\item{only_orf}{Boolean. only return open reading frames.}
}
\description{
get kozak strength (Assumes AUG as start codon)
}
\details{
The kozak sequence hierarchy has been used (Strong to weak):
 
\itemize{
 \item \code{strong_1 = '(A|G)..ATGG'}
 \item \code{moderate_1 =   '(A|G)..ATG.'} 
 \item \code{moderate_2 =   '...ATGG'} 
 \item \code{weak_1 =   '...ATG.'} 
 \item \code{weak_2 = 'ATG'}
}
}
\note{
currently this only check for kozaks in open reading frames.
}
\references{
Whiffin et al, 2020
}
