% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_oorf.R
\name{get_oorf}
\alias{get_oorf}
\title{get overlapping open reading frames}
\usage{
get_oorf(utr, cds, inframe = NULL)
}
\arguments{
\item{utr}{utr sequence}

\item{cds}{cds sequence}

\item{inframe}{boolean should the ORFs be in frame with CDS?}
}
\description{
get overlapping open reading frames
}
