% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_orf.R
\name{get_orf}
\alias{get_orf}
\title{find open reading frames}
\usage{
get_orf(x, start = "ATG", stop = "(TAG)|(TAA)|(TGA)", share_stops = F)
}
\arguments{
\item{x}{a string}

\item{start}{what start codon should be used}

\item{stop}{what stop codons should be used}

\item{share_stops}{boolean. If true (default), then ALL open reading frames are returned. However,
if false, when multiple start codons have the same stop codon, only the one with the strongest kozak 
will be returned.}
}
\description{
takes in a string and finds open reading frames (ORFs)
}
\note{
This function will use the first stoped codon matching a start codon.
The rest of the stop codons (in-frame) will not be considered for the corresponding start codon.
}
