% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_scatter_dens.R
\name{gg_scatter_dens}
\alias{gg_scatter_dens}
\title{ggplot density and scatter}
\usage{
gg_scatter_dens(
  data,
  mapping,
  mapping.dens,
  title = "",
  subtitle = "",
  xlab = "Normalized Log RNA expression",
  ylab = "Normalized Log Protein Expression",
  debug = F
)
}
\description{
ggplot for plotting scatter plot and density map 
for single scat
}
