% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_seq.R
\name{sim_seq}
\alias{sim_seq}
\title{simulate codon frequency}
\usage{
sim_seq(seq, f, k = 2, iter = 1000, verbose = T)
}
\arguments{
\item{seqs}{vector of sequences}

\item{codon}{what codon should be determined}
}
\description{
simulates codon frequency
}
